requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

tortoise-orm==0.17.8
aiosqlite==0.17.0
aiohttp==3.8.1
aiogram==2.16
json5==0.9.6
python-dotenv==0.19.2
envparse==0.2.0
